requirements.txt
fastapi>=0.63.0,<0.64.0
uvicorn>=0.13.4,<0.14.0
databases>=0.4.3,<0.5.0
PyJWT>=2.0.1,<2.1.0
passlib>=1.7.4,<1.8.0
bcrypt>=3.2.0<3.3.0
python-multipart>=0.0.5,<0.1.0
